Finished Gameplay Prototype, now for the tutorial
I added the ChatGPT code to transition to the page adding screen after you guess the right letter, and when you choose a page, it adds it to the book. Added place holder scores for the end screen, and added the new assets Daniel made. Now I need to finish the tutorial